+++ gnoman/security/__init__.py
@@ -1,5 +1,0 @@
-"""Security primitives used by GNOMAN."""
-
-from .encrypted_store import EncryptedJSONStore, EncryptedStoreError
-
-__all__ = ["EncryptedJSONStore", "EncryptedStoreError"]